register-test-user - split _ui_test_helpers into files in _lib
